- Fixed bugs in ANSYS .rst writer. - Added ANSYS 12.0 interface. - Fixed some compile problems on Ubuntu 8.04 - Removed CTest scripts of lse27 and lse35. - Adapted CTest scripts for sedici and ubuntu804 to   cplreader_performance_suite on qnap. - Added latest nightly build shell scripts for rom and sedici.

+++ source/Elements/basefe.cc
@@ -1542,7 +1542,7 @@ namespace CoupledField
 
           std::string str = "4";
           integNode->Get( "order", str );
-          IntegOrder = lexical_cast<Double>(str);
+          IntegOrder = lexical_cast<UInt>(str);
         }
         else
         {
